refactor: ...vo.OrderResponse를 ...dto.OrderDto로 변경
- OrderResponse 역시 지금은 단순히 데이터를 전달하는 용도밖에 없다고 판단함
  - 전체적인 그림에서 봤을 때 어떤 로직이 필요할지는 미지수
  - 추후 필요할 경우 user-service 내에서 별도의 entity를 따로 둔다고 하더라도
    - 지금 현재의 역할은 데이터 전달만으로 충분하다고 생각함
  - 불변성과 값 역할을 하는 vo 취급을 하기에는 애매하다고 판단함
