+++ user-service/src/main/java/com/example/userservice/dto/OrderDto.java
@@ -1,13 +1,12 @@
-package com.example.userservice.vo;
+package com.example.userservice.dto;
 
 import java.time.LocalDateTime;
 
-public record OrderResponse(
+public record OrderDto(
         String productId,
         int quantity,
         int unitPrice,
         int totalPrice,
         LocalDateTime createdAt,
         String orderId
-) {
-}
+) {}
